fix: finalized get public key API
