+++ packages/matrix-identity-server/src/index.test.ts
@@ -1,4 +1,9 @@
-import { Hash, randomString, supportedHashes } from '@twake/crypto'
+import {
+  Hash,
+  randomString,
+  supportedHashes,
+  generateKeyPair
+} from '@twake/crypto'
 import express from 'express'
 import fs from 'fs'
 import fetch from 'node-fetch'
@@ -23,6 +28,8 @@ let idServer: IdServer
 let app: express.Application
 let validToken: string
 let conf: Config
+let longKeyPair: { publicKey: string; privateKey: string; keyId: string }
+let shortKeyPair: { publicKey: string; privateKey: string; keyId: string }
 
 beforeAll((done) => {
   conf = {
@@ -509,6 +516,131 @@ describe('Use configuration file', () => {
           .set('Accept', 'application/json')
         expect(response.statusCode).toBe(401)
       })
+    })
+  })
+
+  describe('Key validation', () => {
+    beforeAll(async () => {
+      // Insert a test key into the database
+      await idServer.db.insert('longTermKeypairs', {
+        keyID: 'testID',
+        public: 'testPub',
+        private: 'testPri'
+      })
+      await idServer.db.insert('shortTermKeypairs', {
+        keyID: 'testID',
+        public: 'testPub',
+        private: 'testPri'
+      })
+    })
+
+    afterAll(async () => {
+      // Remove the test key from the database
+      await idServer.db.deleteEqual('longTermKeypairs', 'keyID', 'testID')
+      await idServer.db.deleteEqual('shortTermKeypairs', 'keyID', 'testID')
+    })
+
+    it('should validate a valid long-term pubkey', async () => {
+      const key = 'testPub'
+      const response = await request(app)
+        .get('/_matrix/identity/v2/pubkey/isvalid')
+        .send({ public_key: key })
+      expect(response.statusCode).toBe(200)
+      expect(response.body.valid).toBe(true)
+    })
+
+    it('should invalidate an invalid long-term pubkey', async () => {
+      const key = 'invalidPub'
+      const response = await request(app)
+        .get('/_matrix/identity/v2/pubkey/isvalid')
+        .send({ public_key: key })
+      expect(response.statusCode).toBe(200)
+      expect(response.body.valid).toBe(false)
+    })
+
+    it('should validate a valid ephemeral pubkey', async () => {
+      const key = 'testPub'
+      const response = await request(app)
+        .get('/_matrix/identity/v2/ephemeral_pubkey/isvalid')
+        .send({ public_key: key })
+      expect(response.statusCode).toBe(200)
+      expect(response.body.valid).toBe(true)
+    })
+
+    it('should invalidate an invalid ephemeral pubkey', async () => {
+      const key = 'invalidPub'
+      const response = await request(app)
+        .get('/_matrix/identity/v2/ephemeral_pubkey/isvalid')
+        .send({ public_key: key })
+      expect(response.statusCode).toBe(200)
+      expect(response.body.valid).toBe(false)
+    })
+  })
+
+  describe('Get public key from keyID', () => {
+    longKeyPair = generateKeyPair('ed25519')
+    shortKeyPair = generateKeyPair('curve25519')
+
+    beforeAll(async () => {
+      // Insert a test key into the database
+      await idServer.db.insert('longTermKeypairs', {
+        keyID: longKeyPair.keyId,
+        public: longKeyPair.publicKey,
+        private: longKeyPair.privateKey
+      })
+      await idServer.db.insert('shortTermKeypairs', {
+        keyID: shortKeyPair.keyId,
+        public: shortKeyPair.publicKey,
+        private: shortKeyPair.privateKey
+      })
+    })
+
+    afterAll(async () => {
+      // Remove the test key from the database
+      await idServer.db.deleteEqual(
+        'longTermKeypairs',
+        'keyID',
+        longKeyPair.keyId
+      )
+      await idServer.db.deleteEqual(
+        'shortTermKeypairs',
+        'keyID',
+        shortKeyPair.keyId
+      )
+    })
+
+    it('should return the public key when correct keyID is given (from long term key pairs)', async () => {
+      const _keyID = longKeyPair.keyId
+      const response = await request(app).get(
+        `/_matrix/identity/v2/pubkey/${_keyID}`
+      )
+
+      expect(response.statusCode).toBe(200)
+      expect(response.body.public_key).toBeDefined()
+      expect(response.body.public_key).toMatch(/^[A-Za-z0-9_-]+$/)
+      expect(response.body.public_key).toBe(longKeyPair.publicKey)
+    })
+
+    it('should return the public key when correct keyID is given (from short term key pairs)', async () => {
+      const _keyID = shortKeyPair.keyId
+      const response = await request(app).get(
+        `/_matrix/identity/v2/pubkey/${_keyID}`
+      )
+
+      expect(response.statusCode).toBe(200)
+      expect(response.body.public_key).toBeDefined()
+      expect(response.body.public_key).toMatch(/^[A-Za-z0-9_-]+$/)
+      expect(response.body.public_key).toBe(shortKeyPair.publicKey)
+    })
+
+    it('should return 404 when incorrect keyID is given', async () => {
+      const _keyID = 'incorrectKeyID'
+      const response = await request(app).get(
+        `/_matrix/identity/v2/pubkey/${_keyID}`
+      ) // exactly '/_matrix/identity/v2/pubkey/' + _keyID
+
+      expect(response.statusCode).toBe(404)
+      expect(response.body.errcode).toBe('M_NOT_FOUND')
     })
   })
 })
