--- packages/matrix-identity-server/src/keyManagement/key.ts
@@ -1,1 +1,0 @@
-// import generateKeyPair from '@twake/crypto'
